Fixed warnings found by Dev14 CTP

+++ Src/ScreenGrab.cpp
@@ -982,7 +982,7 @@ HRESULT DirectX::SaveWICTextureToFile( _In_ ID3D11DeviceContext* pContext,
             return hr;
         }
 
-        WICRect rect = { 0, 0, desc.Width, desc.Height };
+        WICRect rect = { 0, 0, static_cast<INT>( desc.Width ), static_cast<INT>( desc.Height ) };
         hr = frame->WriteSource( FC.Get(), &rect );
         if ( FAILED(hr) )
         {
